requirements.txt
streamlit
pageindex
openai
requests
python-dotenv
